+++ cx-portal/src/components/overlays/AddusersIDP/index.tsx
@@ -32,7 +32,7 @@ import {
   Typography,
 } from 'cx-portal-shared-components'
 import { useDispatch, useSelector } from 'react-redux'
-import { closeOverlay } from 'features/control/overlay/actions'
+import { closeOverlay } from 'features/control/overlay'
 import { useCallback, useEffect, useMemo, useState } from 'react'
 import {
   useFetchIDPDetailQuery,
@@ -46,12 +46,25 @@ import {
   editIDPUserSelector,
   FORMS,
   storeForm,
-} from 'features/control/formSlice'
+} from 'features/control/form'
 import { useDropzone } from 'react-dropzone'
-import IDPStateNotification, {
-  FileFormat,
-  IDPState,
-} from 'components/pages/IDPManagement/IDPStateNotification'
+import { error, success } from 'services/NotifyService'
+
+enum IDPState {
+  SUCCESS_VALID_FORMAT = 'SUCCESS_VALID_FORMAT',
+  SUCCESS_UPLOAD_USERS = 'SUCCESS_UPLOAD_USERS',
+  SUCCESS_DELETE_IDP = 'SUCCESS_DELETE_IDP',
+  ERROR_MULTIPLE_FILES = 'ERROR_MULTIPLE_FILES',
+  ERROR_INVALID_TYPE = 'ERROR_INVALID_TYPE',
+  ERROR_INVALID_FORMAT = 'ERROR_INVALID_FORMAT',
+  ERROR_UPLOAD_USERS = 'ERROR_UPLOAD_USERS',
+  ERROR_DELETE_IDP = 'ERROR_DELETE_IDP',
+}
+
+enum FileFormat {
+  JSON = 'JSON',
+  CSV = 'CSV',
+}
 
 const SelectFormat = ({
   format,
@@ -94,14 +107,14 @@ const AddusersIDPResponse = ({
   return (
     <Dialog open={true}>
       <DialogHeader
-        title={t('users.success.title')}
-        intro={t('users.success.subtitle')}
+        title={t('userssuccess.title')}
+        intro={t('userssuccess.subtitle')}
         closeWithIcon={true}
         onCloseWithIcon={() => storeResponse('')}
       />
       <DialogContent>
         <div>
-          <Typography>{t('users.success.desc')}</Typography>
+          <Typography>{t('userssuccess.desc')}</Typography>
           <pre
             style={{
               padding: '12px',
@@ -130,9 +143,9 @@ export const AddusersIDP = ({ id }: { id: string }) => {
   const [format, setFormat] = useState<FileFormat>(FileFormat.CSV)
   const [pretty, setPretty] = useState<boolean>(false)
   const [unlinked, setUnlinked] = useState<boolean>(false)
+  const [status, setStatus] = useState<boolean | undefined>(undefined)
   const userContent = useSelector(editIDPUserSelector)
   const userResponse = useSelector(editIDPUserResponseSelector)
-  const [status, setStatus] = useState<IDPState>(IDPState.NONE)
   const fetching = t('state.fetching')
 
   const CSV_COLUMNS = useMemo(
@@ -169,12 +182,15 @@ export const AddusersIDP = ({ id }: { id: string }) => {
     )
       .then((response) => response.text())
       .then((result) => {
-        console.log(result)
-        setStatus(IDPState.SUCCESS_UPLOAD_USERS)
         storeResponse(result)
-        setTimeout(() => setStatus(IDPState.NONE), 5000)
+        success(t(`state.${IDPState.SUCCESS_UPLOAD_USERS}`))
+        setStatus(true)
       })
-      .catch((_error) => setStatus(IDPState.ERROR_UPLOAD_USERS))
+      .catch((err) => {
+        error(t(`state.${IDPState.ERROR_UPLOAD_USERS}`, '', err))
+        setStatus(false)
+      })
+    setTimeout(() => setStatus(undefined), 3000)
   }
 
   const downloadUserfile = () => {
@@ -305,17 +321,18 @@ export const AddusersIDP = ({ id }: { id: string }) => {
           })
         )
       } catch (e) {
-        setStatus(IDPState.ERROR_INVALID_FORMAT)
+        error(t(IDPState.ERROR_INVALID_FORMAT))
       }
-      setTimeout(() => setStatus(IDPState.NONE), 5000)
     },
-    [dispatch]
+    [dispatch, t]
   )
 
   const onDrop = useCallback(
     (acceptedFiles: File[]) => {
       if (acceptedFiles.length > 1) {
-        setStatus(IDPState.ERROR_MULTIPLE_FILES)
+        error(t(`state.${IDPState.ERROR_MULTIPLE_FILES}`))
+        setStatus(false)
+        setTimeout(() => setStatus(undefined), 3000)
         return
       }
       const MIME_TYPE = {
@@ -324,7 +341,9 @@ export const AddusersIDP = ({ id }: { id: string }) => {
       }
       acceptedFiles.forEach((file: File) => {
         if (!Object.values(MIME_TYPE).includes(file.type)) {
-          setStatus(IDPState.ERROR_INVALID_TYPE)
+          error(t(`state.${IDPState.ERROR_INVALID_TYPE}`))
+          setStatus(false)
+          setTimeout(() => setStatus(undefined), 3000)
           return
         }
         const reader = new FileReader()
@@ -342,7 +361,7 @@ export const AddusersIDP = ({ id }: { id: string }) => {
         reader.readAsText(file)
       })
     },
-    [csv2json, storeData]
+    [csv2json, storeData, t]
   )
 
   const { getRootProps } = useDropzone({ onDrop })
@@ -406,12 +425,8 @@ export const AddusersIDP = ({ id }: { id: string }) => {
               lineHeight: '20px',
               borderRadius: '24px',
             },
-            ...(status.startsWith('ERROR')
-              ? { backgroundColor: '#FEE7E2' }
-              : {}),
-            ...(status.startsWith('SUCCESS')
-              ? { backgroundColor: '#dfd' }
-              : {}),
+            ...(status === false ? { backgroundColor: '#FEE7E2' } : {}),
+            ...(status === true ? { backgroundColor: '#dfd' } : {}),
           }}
           disabled={true}
           minRows={10}
@@ -436,7 +451,7 @@ export const AddusersIDP = ({ id }: { id: string }) => {
         </Typography>
         <div {...getRootProps()}>
           <DropArea
-            error={status.startsWith('ERROR')}
+            error={status === false}
             size="small"
             translations={{
               title: t('users.drop.title'),
@@ -454,7 +469,6 @@ export const AddusersIDP = ({ id }: { id: string }) => {
           </Typography>
           <Typography>{t('users.drop.note')}</Typography>
         </div>
-        <IDPStateNotification state={status} />
         {userResponse?.data && (
           <AddusersIDPResponse
             response={userResponse.data}