+++ cx-portal/src/components/shared/basic/ReleaseProcess/ValidateAndPublish/index.tsx
@@ -27,6 +27,7 @@ import {
   ConsentStatusEnum,
   useFetchAppStatusQuery,
   useFetchDocumentByIdMutation,
+  useFetchRolesDataQuery,
   useSubmitappMutation,
 } from 'features/appManagement/apiSlice'
 import { setAppStatus } from 'features/appManagement/actions'
@@ -47,6 +48,7 @@ export default function ValidateAndPublish({
   const fetchAppStatus = useFetchAppStatusQuery(appId ?? '', {
     refetchOnMountOrArgChange: true,
   }).data
+  const { data } = useFetchRolesDataQuery(appId ?? '')
 
   useEffect(() => {
     if (fetchAppStatus) dispatch(setAppStatus(fetchAppStatus))
@@ -55,17 +57,6 @@ export default function ValidateAndPublish({
   const defaultValues = useMemo(() => {
     return {
       images: [LogoGrayData, LogoGrayData, LogoGrayData],
-      connectedTableData: {
-        head: ['Linked to your identity', 'Linked NOT to your identity'],
-        body: [
-          ['personal Information', 'Loreum personal Information'],
-          ['Used Content', 'Loreum Used Content'],
-          ['Catena-X Account Data', 'Loreum Catena-X Account Data'],
-        ],
-      },
-      dataSecurityInformation: t(
-        'content.apprelease.defaultValues.dataSecurityInformation'
-      ),
       conformityDocumentsDescription: t(
         'content.apprelease.defaultValues.conformityDocumentsDescription'
       ),
@@ -117,12 +108,8 @@ export default function ValidateAndPublish({
         longDescriptionTitleDE={t(
           'content.apprelease.validateAndPublish.longDescriptionTitleDE'
         )}
-        connectedData={t('content.apprelease.validateAndPublish.connectedData')}
         conformityDocument={t(
           'content.apprelease.validateAndPublish.conformityDocument'
-        )}
-        dataSecurityInformation={t(
-          'content.apprelease.validateAndPublish.dataSecurityInformation'
         )}
         documentsTitle={t('content.apprelease.validateAndPublish.documents')}
         providerInformation={t(
@@ -137,6 +124,7 @@ export default function ValidateAndPublish({
         helpText={t('content.apprelease.footerButtons.help')}
         submitButton={t('content.apprelease.footerButtons.submit')}
         values={defaultValues}
+        rolesData={data}
       />
     </div>
   )