+++ cx-portal/src/features/apps/apiSlice.ts
@@ -57,8 +57,6 @@ export enum SubscriptionStatus {
   ACTIVE = 'ACTIVE',
   PENDING = 'PENDING',
   INACTIVE = 'INACTIVE',
-  IN_REVIEW = 'IN_REVIEW',
-  CREATED = 'CREATED',
 }
 
 export enum SubscriptionStatusText {
@@ -72,6 +70,8 @@ export enum SubscriptionStatusText {
 export type SubscriptionStatusItem = {
   appId: string
   offerSubscriptionStatus: SubscriptionStatus
+  name: string
+  provider: string
 }
 
 export enum DocumentTypeText {
@@ -98,6 +98,11 @@ export type AppDetails = AppMarketplaceApp & {
   leadPictureUri?: ImageType
   privacyPolicies: PrivacyPolicyType[]
   roles?: string[]
+  description?: {
+    languageCode: string
+    longDescription: string
+    shortDescription: string
+  }[]
 }
 
 export type Documents = {
@@ -133,6 +138,17 @@ export type DocumentRequestData = {
   documentId: string
 }
 
+export type ActiveAppsData = {
+  id: string
+  name: string
+  shortDescription: string
+  provider: string
+  price: string
+  leadPictureId: string
+  useCases: string[]
+  status?: string
+}
+
 export const apiSlice = createApi({
   reducerPath: 'rtk/apps/marketplace',
   baseQuery: fetchBaseQuery(apiBaseQuery()),
@@ -141,7 +157,25 @@ export const apiSlice = createApi({
       query: (id: string) => `/api/apps/${id}?lang=${i18next.language}`,
     }),
     fetchActiveApps: builder.query<AppMarketplaceApp[], void>({
-      query: () => `/api/apps/active`,
+      async queryFn(_arg, _queryApi, _extraOptions, fetchWithBQ) {
+        const activeApps = await fetchWithBQ(`/api/apps/active`)
+        if (activeApps.error) return { error: activeApps.error }
+        const data = activeApps.data as AppMarketplaceApp[]
+        const subscriptionStatus = await fetchWithBQ(
+          `/api/apps/subscribed/subscription-status`
+        )
+        const subscriptionData =
+          subscriptionStatus.data as SubscriptionStatusItem[]
+        data.forEach(async (appItem: AppMarketplaceApp) => {
+          subscriptionData.forEach(
+            async (subscriptionItem: SubscriptionStatusItem) => {
+              if (appItem.id === subscriptionItem.appId)
+                appItem.status = subscriptionItem.offerSubscriptionStatus
+            }
+          )
+        })
+        return { data: data }
+      },
     }),
     fetchFavoriteApps: builder.query<string[], void>({
       query: () => `/api/apps/favourites`,